Refactored merging code used in tests.
